backend/requirements/base.txt
asgiref==3.2.3
Django==3.0.1
django-autoslug==1.9.6
django-filter==2.2.0
django-guardian==2.1.0
django-js-asset==1.2.2
django-model-utils==4.0.0
django-mptt==0.10.0
django-teryt-tree==0.15.1
djangorestframework==3.11.0
lxml==4.4.2
Markdown==3.1.1
mysqlclient==1.4.6
pytz==2019.3
sqlparse==0.3.0
tqdm==4.41.1
